Semifinal/project/server（云端服务）/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
sqlalchemy[asyncio]>=2.0.30
aiosqlite>=0.20.0
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0
